include all students when updating overrides
fixes OUT-518

test plan:
 1. Select or create a course with three students.
 2. Create an assignment.
 3. Create an override for two of students.
 4. Deactivate one of the students from the override.
 5. Visit the Edit Assignment page for this assignment.
 6. Change the date of the override.
 7. Save the changes.
 8. Verify the date was changed successfully.

Reviewed-on: https://gerrit.instructure.com/125590
Tested-by: Jenkins
QA-Review: Andrew Porter <hporter-c@instructure.com>
Product-Review: Keith T. Garner <kgarner@instructure.com>

+++ lib/api/v1/assignment_override.rb
@@ -127,10 +127,10 @@ module Api::V1::AssignmentOverride
         errors << "invalid student_ids #{student_ids.inspect}"
         students = []
       else
-        # look up all the active students since the assignment will affect all
-        # active students in the course on this override and not just what the
+        # look up all students since the assignment will affect all current and
+        # previous students in the course on this override and not just what the
         # teacher can see that were sent in the request object
-        students = api_find_all(assignment.context.students.active, student_ids)
+        students = api_find_all(assignment.context.all_students, student_ids)
         students = students.distinct if students.is_a?(ActiveRecord::Relation)
         students = students.uniq if students.is_a?(Array)
 
@@ -426,6 +426,8 @@ module Api::V1::AssignmentOverride
     prepared_overrides[:override_errors].each do |error|
       assignment.errors.add(:base, error)
     end
+
+    raise ActiveRecord::RecordInvalid.new(assignment) if assignment.errors.any?
 
     if prepared_overrides[:overrides_to_delete].any?
       assignment.assignment_overrides.where(id: prepared_overrides[:overrides_to_delete]).destroy_all
